Binary format user headers and YAML binary support.
User headers in the get_data binary format are now
interpreted correctly and can be retrieved with
a convenience method.

YAML binaries have a single YAML document as
header, followed by raw binary data, similar
to "the binary" format. Note that YAML binaries
are not supported by PyPy.

So, duh, it was not a smart idea to implement it in
the first place.

And so the code is not tested.

Have fun with it =)

git-svn-id: https://devel-ik.fzk.de/svn/lhc2cr/sipm/projects/Gregor@1192 fd7b642b-f7a7-476b-83dd-42b6f92cd3b3