Added "full-spectrum" output and converted plot-scripts
If the "full-spectrum" cmdline flag is set, all detected pulses in the
spectrum are written to the output stream. The detection routine is the
same as in extract_countrate, but the output format is differently
(not only the counts/frame, but each pulse with its amplitude separately).

Introducing this flag, extract_countrate is obsolete. Further analysis
use the pulse spectrum CSV layout as input, now. In the CSV header,
an entry with the spectrum type is written, so an analysis script can
deduce whether the it is suitable for its kind of analysis.

git-svn-id: https://devel-ik.fzk.de/svn/lhc2cr/sipm/projects/Gregor@1195 fd7b642b-f7a7-476b-83dd-42b6f92cd3b3